@@ -1,49 +1,51 @@
 {
-  "name": "church-dinner-finder",
+  "name": "x",
   "version": "0.0.0",
   "license": "MIT",
-  "angular-cli": {},
   "scripts": {
+    "ng": "ng",
     "start": "ng serve",
-    "lint": "tslint \"src/**/*.ts\"",
+    "build": "ng build",
     "test": "ng test --watch=false",
-    "pree2e": "webdriver-manager update",
-    "e2e": "protractor"
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "2.2.1",
-    "@angular/compiler": "2.2.1",
-    "@angular/core": "2.2.1",
-    "@angular/forms": "2.2.1",
-    "@angular/http": "2.2.1",
-    "@angular/platform-browser": "2.2.1",
-    "@angular/platform-browser-dynamic": "2.2.1",
-    "@angular/router": "3.2.1",
-    "@ng-bootstrap/ng-bootstrap": "1.0.0-alpha.11",
-    "bootstrap": "4.0.0-alpha.5",
-    "core-js": "^2.4.1",
-    "rxjs": "5.0.0-beta.12",
-    "ts-helpers": "^1.1.1",
-    "zone.js": "^0.6.23"
-  },
-  "devDependencies": {
-    "@angular/compiler-cli": "2.2.1",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "^6.0.42",
-    "angular-cli": "1.0.0-beta.21",
-    "codelyzer": "~1.0.0-beta.3",
-    "jasmine-core": "2.5.2",
-    "jasmine-spec-reporter": "2.5.0",
-    "karma": "1.2.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-cli": "^1.0.1",
-    "karma-jasmine": "^1.0.2",
-    "karma-remap-istanbul": "^0.2.1",
-    "protractor": "4.0.9",
-    "ts-node": "1.2.1",
-    "tslint": "3.13.0",
-    "typescript": "~2.0.3",
-    "webdriver-manager": "10.2.5"
-  }
+  "@angular/animations": "^5.0.2",
+  "@angular/common": "^5.0.2",
+  "@angular/compiler": "^5.0.2",
+  "@angular/core": "^5.0.2",
+  "@angular/forms": "^5.0.2",
+  "@angular/http": "^5.0.2",
+  "@angular/platform-browser": "^5.0.2",
+  "@angular/platform-browser-dynamic": "^5.0.2",
+  "@angular/router": "^5.0.2",
+  "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.5",
+  "bootstrap": "4.0.0-alpha.6",
+  "core-js": "^2.4.1",
+  "rxjs": "^5.5.2",
+  "zone.js": "^0.8.14"
+},
+"devDependencies": {
+  "@angular/cli": "1.5.0",
+  "@angular/compiler-cli": "^5.0.2",
+  "@angular/language-service": "^5.0.2",
+  "@types/jasmine": "~2.5.53",
+  "@types/jasminewd2": "~2.0.2",
+  "@types/node": "~6.0.60",
+  "codelyzer": "~3.2.0",
+  "jasmine-core": "~2.6.2",
+  "jasmine-spec-reporter": "~4.1.0",
+  "karma": "~1.7.0",
+  "karma-chrome-launcher": "~2.1.1",
+  "karma-cli": "~1.0.1",
+  "karma-coverage-istanbul-reporter": "^1.2.1",
+  "karma-jasmine": "~1.1.0",
+  "karma-jasmine-html-reporter": "^0.2.2",
+  "protractor": "~5.1.2",
+  "ts-node": "~3.2.0",
+  "tslint": "~5.7.0",
+  "typescript": "~2.4.2"
+}
 }
