@@ -1,9 +1,12 @@
 import {Injectable} from '@angular/core';
 import {Http, Response} from '@angular/http';
 
-import 'rxjs/add/operator/cache';
+import 'rxjs/add/operator/publishReplay';
 import 'rxjs/add/operator/map';
-import {Observable} from 'rxjs';
+import {Observable} from 'rxjs/Observable';
+
+import { environment } from '../../environments/environment';
+
 
 @Injectable()
 export class NavService {
@@ -14,9 +17,10 @@ export class NavService {
 
   getNavData(): Observable<any> {
     if (!this._navData) {
-      this._navData = this.http.get('http://localhost:8080/application')
+      this._navData = this.http.get(environment.serverUrl + 'application')
           .map((res: Response) => res.json())
-          .cache();
+          .publishReplay()
+          .refCount();
     }
     return this._navData;
   }
