@@ -32,8 +32,5 @@ if (Environment.isDevelopmentMode() && targetDir != null) {
         }
     }
     logger("StackTrace", ERROR, ['FULL_STACKTRACE'], false)
-    root(ERROR, ['STDOUT', 'FULL_STACKTRACE'])
 }
-else {
-    root(ERROR, ['STDOUT'])
-}
+root(ERROR, ['STDOUT'])
